# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Sistema de Gestión de Subvenciones API
  description: |
    API REST para el Sistema de Gestión de Subvenciones de Alquiler.
    
    ## Autenticación
    La API utiliza JWT (JSON Web Tokens) para autenticación. Los tokens se envían en el header `Authorization` con el prefijo `Bearer`.
    
    ## Roles
    - **BENEFICIARIO**: Solicitante de la subvención
    - **ARRENDADOR**: Propietario del inmueble
    - **DIGER**: Personal administrativo que gestiona solicitudes
    - **DIRECTORA**: Directora que aprueba solicitudes
    - **ORDENADOR_GASTO**: Responsable de firma final
    - **CRI**: Centro de Recursos Informáticos (auditoría)
    
    ## Estados del Proceso
    - BORRADOR → ENVIADA → EN_VALIDACION → PENDIENTE_DIRECTORA → APROBADA_DIRECTORA → REVISION_ORDENADOR → APROBADA → CERRADA
    - También: RECHAZADA, CORRECCION_SOLICITADA
  version: 1.0.0
  contact:
    name: Soporte Técnico
    email: soporte@subvenciones.gob.es
  license:
    name: Propietario
servers:
  - url: http://localhost:4000/api
    description: Servidor de desarrollo
  - url: https://api.subvenciones.gob.es/api
    description: Servidor de producción
...
paths:
...
  /auth/login:
    post:
      tags: [Autenticación]
      summary: Iniciar sesión
      description: Autentica un usuario y devuelve tokens JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cedula, password]
              properties:
                cedula:
                  type: string
                  example: "12345678-9"
                password:
                  type: string
                  format: password
                  example: "MiContraseña123!"
      responses:
        '200':
          description: Login exitoso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '202':
          description: Requiere MFA
          content:
            application/json:
              schema:
                type: object
                properties:
                  requiresMfa:
                    type: boolean
                    example: true
                  tempToken:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '423':
          description: Cuenta bloqueada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /auth/refresh:
    post:
      tags: [Autenticación]
      summary: Renovar token de acceso
      description: Usa el refresh token (cookie httpOnly) para obtener nuevo access token
      responses:
        '200':
          description: Token renovado
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags: [Autenticación]
      summary: Cerrar sesión
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Sesión cerrada
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Sesión cerrada correctamente"
  /auth/me:
    get:
      tags: [Autenticación]
      summary: Obtener usuario actual
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Información del usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/change-password:
    post:
      tags: [Autenticación]
      summary: Cambiar contraseña
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword:
                  type: string
                newPassword:
                  type: string
                  minLength: 10
      responses:
        '200':
          description: Contraseña cambiada
        '400':
          $ref: '#/components/responses/BadRequest'
  /auth/mfa/setup:
    post:
      tags: [Autenticación]
      summary: Configurar MFA
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Datos para configurar MFA
          content:
            application/json:
              schema:
                type: object
                properties:
                  secret:
                    type: string
                  qrCode:
                    type: string
                    description: Data URL de imagen QR
  /auth/mfa/verify:
    post:
      tags: [Autenticación]
      summary: Verificar y activar MFA
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
      responses:
        '200':
          description: MFA activado

  # ==================== USUARIOS ====================
...
  /auditoria/eventos:
    get:
      tags: [Auditoría]
      summary: Obtener eventos de auditoría
      description: Solo roles DIGER y CRI
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - name: accion
          in: query
          schema:
            type: string
        - name: usuarioId
          in: query
          schema:
            type: string
            format: uuid
        - name: desde
          in: query
          schema:
            type: string
            format: date-time
        - name: hasta
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Lista de eventos
          content:
            application/json:
              schema:
                type: object
                properties:
                  eventos:
                    type: array
                    items:
                      $ref: '#/components/schemas/EventoAuditoria'
                  total:
                    type: integer
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /auditoria/estadisticas:
    get:
      tags: [Auditoría]
      summary: Obtener estadísticas de auditoría
      description: Solo rol CRI
      security:
        - bearerAuth: []
      parameters:
        - name: desde
          in: query
          schema:
            type: string
            format: date
        - name: hasta
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Estadísticas
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalEventos:
                    type: integer
                  eventosPorAccion:
                    type: object
                  eventosPorDia:
                    type: array
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
...
  responses:
    BadRequest:
      description: Solicitud inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: No autenticado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Sin permisos
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
    Rol:
      type: string
      enum: [BENEFICIARIO, ARRENDADOR, DIGER, DIRECTORA, ORDENADOR_GASTO, CRI]
    EstadoUsuario:
      type: string
      enum: [ACTIVO, INACTIVO, BLOQUEADO]
...
    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        cedula:
          type: string
        email:
          type: string
          format: email
        nombreCompleto:
          type: string
        rol:
          $ref: '#/components/schemas/Rol'
        estado:
          $ref: '#/components/schemas/EstadoUsuario'
        mfaHabilitado:
          type: boolean
        creadoEn:
          type: string
          format: date-time
...
    EventoAuditoria:
      type: object
      properties:
        id:
          type: string
          format: uuid
        accion:
          type: string
        entidad:
          type: string
        entidadId:
          type: string
        detalles:
          type: object
        ip:
          type: string
        userAgent:
          type: string
        usuario:
          type: object
          properties:
            id:
              type: string
            nombreCompleto:
              type: string
        fecha:
          type: string
          format: date-time
security:
  - bearerAuth: []